BreakageClassifier/code/requirements.txt
pandas
openwpm-utils==0.2.0
tqdm==4.64.0
pyyaml==6.0
resource==0.2.1
multidict==6.0.2
adblockparser==0.7
plyvel
scikit-learn==1.1.3
scipy==1.8.1
tldextract==2.2.2
urllib3==1.26.9
Levenshtein==0.18.1
networkx==2.8.4
leveldb
treeinterpreter==0.2.3
aiofiles==0.8.0
matplotlib
nltk
python-dotenv
seaborn
validators
langdetect
imbalanced-learn
plotly==5.18.0